+++ src/HasManipulators.php
@@ -87,7 +87,7 @@ trait HasManipulators
             $property = Str::camel($property);
             if (property_exists($this, $property)) {
                 $this->$property = request($key);
-            } else if (property_exists($this, Str::camel($key))) {
+            } elseif (property_exists($this, Str::camel($key))) {
                 $this->{Str::camel($key)} = request($key);
             }
         }
@@ -97,9 +97,11 @@ trait HasManipulators
 
     public function setInitialState($state)
     {
-        if (is_object($state)) {
+        if (is_object($state) && (!$state instanceof Closure)) {
             $this->initialState = get_class($state);
-        } else if (is_string($state) && class_exists($state)) {
+        } elseif (is_string($state) && class_exists($state)) {
+            $this->initialState = $state;
+        } elseif (is_callable($state)) {
             $this->initialState = $state;
         } else {
             $this->initialState = null;