+++ tests/StateTest.php
@@ -3,11 +3,15 @@
 namespace Sparors\Ussd\Tests;
 
 use Orchestra\Testbench\TestCase;
+use Sparors\Ussd\State;
 
 class StateTest extends TestCase
 {
-    public function testBasic()
+    public function testState()
     {
-        $this->assertTrue(true);
+        $hello = new HelloState();
+        $this->assertEquals('Hello World', $hello->render());
+
+        $this->assertEquals(ByeState::class, $hello->next('1'));
     }
 }